refactor(fullduplex): remove duplicate duplex runtime
