+++ vllm_omni/entrypoints/async_omni.py
@@ -35,7 +35,7 @@ from vllm_omni.entrypoints.omni_base import (
     OmniEngineDeadError,
 )
 from vllm_omni.errors import client_error_metadata
-from vllm_omni.experimental.duplex.engine import duplex_data_plane_request_info
+from vllm_omni.experimental.fullduplex.engine.omni import duplex_data_plane_request_info
 from vllm_omni.inputs.data import OmniSamplingParams
 from vllm_omni.metrics.stats import OrchestratorAggregator as OrchestratorMetrics
 from vllm_omni.outputs import OmniRequestOutput
