--- vllm_omni/experimental/duplex/__init__.py
@@ -1,1 +1,0 @@
-"""Experimental session-oriented MiniCPM duplex runtime."""
