--- vllm_omni/experimental/duplex/engine.py
@@ -1,4 +1,0 @@
-"""Deprecated compatibility imports for the consolidated duplex engine port."""
-
-from vllm_omni.experimental.fullduplex.engine.omni import *  # noqa: F403
-from vllm_omni.experimental.fullduplex.engine.omni import __all__  # noqa: F401
