--- vllm_omni/experimental/duplex/models/__init__.py
@@ -1,1 +1,0 @@
-"""Model-specific implementations for the experimental duplex runtime."""
