--- vllm_omni/experimental/duplex/models/minicpmo45/__init__.py
@@ -1,1 +1,0 @@
-"""MiniCPM-o 4.5 experimental native-duplex implementation."""
