--- vllm_omni/experimental/duplex/models/minicpmo45/policy.py
@@ -1,5 +1,0 @@
-"""Compatibility import for the consolidated MiniCPM policy."""
-
-from vllm_omni.experimental.fullduplex.minicpmo45.policy import MiniCPMO45DuplexPolicy
-
-__all__ = ["MiniCPMO45DuplexPolicy"]
