--- vllm_omni/experimental/duplex/openai/__init__.py
@@ -1,1 +1,0 @@
-"""OpenAI-compatible adapters for the experimental duplex runtime."""
