--- vllm_omni/experimental/duplex/openai/adapters/__init__.py
@@ -1,6 +1,0 @@
-from vllm_omni.experimental.duplex.openai.adapters.minicpmo45 import (
-    MiniCPMO45NativeDuplexServingAdapter,
-    MiniCPMO45PcmAppendBuffer,
-)
-
-__all__ = ["MiniCPMO45NativeDuplexServingAdapter", "MiniCPMO45PcmAppendBuffer"]
